+++ src/main/java/de/catma/ui/module/annotate/CommentMessageListener.java
@@ -58,13 +58,13 @@ public class CommentMessageListener extends UIMessageListener<CommentMessage> {
 			CommentMessage commentMessage = message.getMessageObject();
 			final String documentId = commentMessage.getDocumentId();
 			final boolean replyMessage = commentMessage.isReplyMessage();
-			final int senderId = commentMessage.getSenderId();
+			final Long senderId = commentMessage.getSenderId();
 			final boolean deleted= commentMessage.isDeleted();
 			final SourceDocumentReference document = documentSupplier.get();
 			if ((document != null) 
 					&& document.getUuid().equals(documentId)) {
 				User user = project.getUser();
-				Integer receiverId = user.getUserId();
+				Long receiverId = user.getUserId();
 				
 				if (!receiverId.equals(senderId)) {
 					final Comment comment = commentMessage.toComment();
